Fixed handle_invitation_request_accept() so it looks for the correct message string.
Also updated the accept and reject handlers so that they call the
on_selected_cb to update the state of the buttons.  This was needed for
when an invitation in the out_inv_tree was already selected and we
received an updated accept/deny message.  Without this change, the
Remove button wouldn't get enabled.


git-svn-id: svn://svn.code.sf.net/p/simias/code/trunk@3466 1a4d904c-e006-0410-be0b-b506c45ff692