Fixing bug #72422:
nautilus-ifolder no longer has to build gSOAP into it.  It uses libifolder
to get the gSOAP calls.

I've also added the ability to start supporting multiple domains when
converting a folder to an iFolder.  This still needs more work (namely,
to change the disabled "Path" entry box to just a label and also give the
user an error message if no domains exist).


git-svn-id: svn://svn.code.sf.net/p/simias/code/trunk@3717 1a4d904c-e006-0410-be0b-b506c45ff692