1. Fixed bug #80906. Reworked invitation code and POBoxService web service. This change will require a server and client upgrade because of the change to the web service. There are no database changes required so the store does not have to be deleted. This bug was caused by several race conditions in the invitation code which was caused when a single iFolder invitation was accepted/declined (or any combination) between two or more clients logged into the same iFolder account. The following race conditions were fixed:
1. Accepting an invitation at the same time from more than one client now results in the iFolder being synced to each machine properly and the subscriptions going to the ready state.
2. Accepting an invitation and declining the invitation at the same time now results in the following:
If the accept was synced first, then the decline is not allowed and the subscription returns back to the ready state on the declined client and is set to the ready state and the iFolder is synced to the accepted client.
If the decline was synced first, then the subscription is removed from all clients.
3. Declining an invitation at the same time now results in the subscription being removed from all clients.

The files and current revisions that changed for this fix are the following:
GlobalProperties.cs rev 1.132
Collection.cs rev 1.123
Collision.cs rev 1.11
Node.cs rev 1.58
POBoxService.asmx.cs rev 1.67
Message.cs rev 1.13
POBox.cs rev 1.37
POBoxManager.cs rev 1.13
POBoxServiceProxy.cs rev 1.15
Subscription.cs rev 1.46
SubscriptionThread.cs rev 1.65
PropertyTags.cs rev 1.37


git-svn-id: svn://svn.code.sf.net/p/simias/code/trunk@4523 1a4d904c-e006-0410-be0b-b506c45ff692