Fix for bug #81540 - libsimias-event leaks memory.
We were incorrectly freeing up memory after making a call to xmlXPathEvalExpression().  The object that is returned has to be freed using xmlXPathFreeObject() and we were using xmlFree().Along with this fix, I also fixed the problem of this not being able to be run from multiple processes.  The library was using IPC semaphores which were tied to a file semaphore.  I replaced this with pthread_cond_*() and pthread_mutex_*() calls.


git-svn-id: svn://svn.code.sf.net/p/simias/code/trunk/simias@4524 1a4d904c-e006-0410-be0b-b506c45ff692