Fix for Bug 119785 - Help doesn't work on NLD SP2 or SUSE 10.
Something in the way that we configure our environment in our /opt/novell/ifolder3/bin/ifolder startup script is preventing Gnome.Url.Show() from working.  So, instead, I modified our code to just spawn the "gnome-open" program, which just invokes gnome_url_show().  If we ever get time to reinvestigate this, we can, but for now, this solves the problem.

Changed file:
	/src/LinuxClient/library/Util.cs.in


git-svn-id: svn://svn.code.sf.net/p/simias/code/trunk@5330 1a4d904c-e006-0410-be0b-b506c45ff692