Additional fix for Bug 128523 - Accounts Dialogue should warn user when they navigate away from page.
The selection changed event handler was being raised before the entries were cleared out on a removal.  So, to fix the problem, I cleared out the entries so that by the time the selection changed handler is invoked, it doesn't think it's in New Account Mode.

File changed (in 3.2 Branch and CVS Head):
	src/LinuxClient/library/PrefsAccountsPage.cs


git-svn-id: svn://svn.code.sf.net/p/simias/code/branches/R3OESSP2RC1_BR@5436 1a4d904c-e006-0410-be0b-b506c45ff692