1. removed stale core directory
2. replayed Russ'es
#6126
fixed Bug 160699 - iFolder 3.4 Simias Process unloads when attempting to sync directory with case conflict

The call to readlink changed and was causing an exception because the StringBuilder had a capacity of 0;

#6127
fixed Bug 156660 - iFolder status shows incorrect "ok" status when server is returning "server is busy" status

The status will now show disconnected. There isn't time to add new strings for the release.

#6128
Added code to help with invalid conflicts.

git-svn-id: svn://svn.code.sf.net/p/simias/code/trunk/simias@6134 1a4d904c-e006-0410-be0b-b506c45ff692