updated configure to default simiasdatadir to /var/lib/simias unless runasclient option is used
Updated the sysconfdir to default to $prefix/etc and append simias in
our configure scripts.  Changed all make files that reference sysconfdir
and also code so it will locate configuration properly.

Added a wrapper to the SimiasSetup.cs class so when setting the prefix
the changes only happen on windows




git-svn-id: svn://svn.code.sf.net/p/simias/code/trunk@6243 1a4d904c-e006-0410-be0b-b506c45ff692